backend/requirements.txt
Django>=3.2,<4.0
django-ninja
instructor
openai
psycopg2-binary